tests/wpt/web-platform-tests/client-hints/accept_ch.tentative.https.html

<html>
<title>Accept-CH test</title>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<!--
Apart from this webpage, the test opens two more html web page. One test is run
in this web page, and two in the other web pages.
-->

<script>

// This test fetches resources/accept_ch.html. The response headers to
// that webpage contains only the Accept-CH header. Due to the missing
// Accept-CH-Lifetime header, the user-agent should not persist origin
// preferences for the client hints specified in Accept-CH header.

// Next, to verify that the origin preferences were not persisted by the user
// agent, this test fetches resources/do_not_expect_client_hints_headers.html
// in a new window. Fetching of
// resources/do_not_expect_client_hints_headers.html
// verifies that the user agent does not send the client hints in the request
// headers.

// Test is marked as tentative until https://github.com/whatwg/fetch/issues/726
// is resolved.

// First, verify the initial state to make sure that the browser does not have
// client hints preferences cached from a previous run of the test.
promise_test(t => {
  return fetch("echo_client_hints_received.py").then(r => {
    assert_equals(r.status, 200)
    // Verify that the browser did not include client hints in the request
    // headers when fetching echo_client_hints_received.py.
    assert_false(r.headers.has("device-memory-received"), "device-memory-received");
  });
}, "Precondition: Test that the browser does not have client hints preferences cached");

async_test(t => {
  window.addEventListener('message', function(e) {
    if(!e.source.location.pathname.includes("do_not_expect_client_hints_headers.html")) {
      return;
    }
    if(typeof e.data != "string")
      return;
    assert_equals(e.data, "PASS");
    t.done();
  })
}, "Loading of resources/do_not_expect_client_hints_headers.html did not finish.");

function acceptChLoaded() {
  // Open a new window. Verify that the user agent does not attach the client
  // hints.
  var verify_win = window.open("resources/do_not_expect_client_hints_headers.html");
  assert_not_equals(verify_win, null, "Popup windows not allowed?");
}

// Fetching this webpage should NOT cause user-agent to persist client hint
// preferences for the origin.
var win = window.open("resources/accept_ch.html");
assert_not_equals(win, null, "Popup windows not allowed?");
win.addEventListener('load', acceptChLoaded, false);

</script>
</body>
</html>
